refactor: improve type imports
