+++ packages/log/ts/memory/index.ts
@@ -1,7 +1,7 @@
 import { required } from 'type-plus'
 import { formatLogLevel, plainLogFormatter } from '../formatter/index.js'
 import { toInspectLogEntry } from '../platform/index.js'
-import { LogEntry, LogFilter, LogFormatter, LogReporter, LogReporterOptions } from '../types.js'
+import type { LogEntry, LogFilter, LogFormatter, LogReporter, LogReporterOptions } from '../types.js'
 import { assertLogModeIsNotProduction } from '../utils.js'
 
 export function logEntriesToString(logs: LogEntry[]) {
